+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './array/random';
+export * from './array/shuffle';